requirements.txt
argcomplete==3.1.1
asgiref==3.7.2
boto3==1.26.157
botocore==1.29.157
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==4.0.0
Django==4.2.2
django-allauth==0.52.0
django-apps==0.1.0
django-config-url==0.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-urls==1.1.3
djangorestframework==3.14.0
durationpy==0.5
filelock==3.12.2
gunicorn==20.1.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kappa==0.6.0
MarkupSafe==2.1.3
numpy==1.25.0
oauthlib==3.2.2
pandas==2.0.2
Pillow==9.5.0
placebo==0.9.0
platformdirs==3.7.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
Werkzeug==2.3.6
zappa==0.57.0
